Refactor to have StudyHome as its own component
